# Parallel encryption