Two days ago I had created a new module in crates.io called "asynchronous". It allows parallel and deferred execution of threads in a easy and safe way.
By now, I modified the encryption part of the project to adapt to this module. Now the encryption process runs in all the CPU cores.

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/dirvine/self_encryption/59)
<!-- Reviewable:end -->