@benjaminbollen 2015-05-11
Ran a benchmark on my laptop; of this PR and maidsafe/async branch both merged into master: there is an overall few MB/s increase in speed; but more importantly: very notably the stress is no longer on a single core, but spread over the cores.  So very nice work.  I have not dived into the code yet.

with async: 
running 5 tests
test bench_write_then_read_a_200B ... bench:     13067 ns/iter (+/- 938) = 30 MB/s
test bench_write_then_read_b_1KB  ... bench:     47447 ns/iter (+/- 1313) = 43 MB/s
test bench_write_then_read_c_1MB  ... bench:  53708306 ns/iter (+/- 1767000) = 37 MB/s
test bench_write_then_read_d_3MB  ... bench: 157556196 ns/iter (+/- 3227794) = 37 MB/s
test bench_write_then_read_e_10MB ... bench: 523736052 ns/iter (+/- 21033036) = 20 MB/s

master (without async):
running 5 tests
test bench_write_then_read_a_200B ... bench:     13838 ns/iter (+/- 982) = 28 MB/s
test bench_write_then_read_b_1KB  ... bench:     49494 ns/iter (+/- 1936) = 41 MB/s
test bench_write_then_read_c_1MB  ... bench:  56313256 ns/iter (+/- 843919) = 35 MB/s
test bench_write_then_read_d_3MB  ... bench: 170048836 ns/iter (+/- 4660920) = 31 MB/s
test bench_write_then_read_e_10MB ... bench: 588400238 ns/iter (+/- 48747987) = 20 MB/s