@bcndanos 2015-05-11
I suppose you have modified the bench.rs for include the Storage struct as "Arc" and "Mutex". I forgot it in the PR.
In that case, the part where the tests "put" to storage don't run in parallel, the threads have a mutual exclusión.
The implementation of Storage Trait in the example/basic_encryptor.rs is even better, aprox. 40%, due to write to disk in parallel.