@benjaminbollen 2015-05-11
The basic_encryptor example indeed gives a notable speed up on the disk IO.  for a 185MB file with the async write out, it shaves off almost 1 minute down to 2min28 on my machine.  So very nice indeed!  What we need next is the implementation to asynchronously SE, because that process still happens synchronously.