@BSMaidSafe 2015-05-14
Hi, Having updated the bench tests as you suggested there's a noticeable performance improvement there also. I'll put in a PR for that, and recommend to David that both can be merged. Thanks again for your contribution. Out of curiosity, do you have any plans for extending, or otherwise maintaining the asynchronous crate?