@bcndanos 2015-05-14
Yes, I'm improving the asynchronous crate continuosly. 
Just now I updated to versión 0.4.1 that allows EventLoops. I'm testing this change with Crust module, in beacon.rs and works perfectly seeking peers. In this case, it doesn't bring any benefit in terms of computational time, but reduces the code significantly.
If you need any change, just PR any change to me.