@BSMaidSafe 2015-05-14
Excellent, good to know. Will do on the PR's if required.