@dirvine 2015-05-14
Very nice work, thank you very much. I See this being very useful in crust, not to help beacon so much but for parallel attempts (up to 15 at a time) to bootstrap from the bootstrap file. This is sorely missing yet and on the roadmap, if you fancy trying that one. Then @chandraprakash will buy you a huge beer.