@bcndanos 2015-05-14
I've been reviewing at the code and I think it's easy to do it. But before I have to develop a new method in my module for processing in parallel until at least one task(in this case one connection) is done.